requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
python-dotenv==1.1.1
httpx==0.27.2
tenacity==9.0.0
pydantic==2.9.2
prometheus-client==0.21.0
phonenumbers==8.13.48
